+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.4.88 (2024-10-21)
+
+- fix: agent view crash
+
+## 0.4.87 (2024-10-20)
+
+- feat: add image type for agent inputs
+- chore: auto manage image blender inputs
+
 ## 0.4.85 (2024-10-19)
 
 - feat: support global config secrets for resource blocklet